% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance-dependent.R
\name{dd}
\alias{dd}
\title{Distance dependent index}
\arguments{
\item{id}{tree identification}

\item{x, }{x numeric vector of tree coordinates.}

\item{dbh}{diameter of trees at 1.3 meters.}

\item{search}{method to search competitive trees. Use "nearest" to find
the \code{n} trees nearest of object tree or "dfixed" to find any tree around
of \code{n} distance.}
}
\description{
These are functions to compute some distante dependente indexs.
each index is named that author who propose.
}
