% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_distance_independent.R
\name{di_hamilton}
\alias{di_hamilton}
\title{Distance-independent Hamilton (1986) index}
\usage{
di_hamilton(dbh)
}
\description{
Ratio of the diameter of the subject tree to the quadratic mean diameter
}

