% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_distance_independent.R
\name{di_lorimer}
\alias{di_lorimer}
\title{Distance-independent Lorimer (1982) index}
\usage{
di_lorimer(dbh)
}
\description{
Sums up the d of neighbours divided by the subject
tree d in the plot (ha-1)
}

