% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance-independent.R
\name{di_vanclay}
\alias{di_vanclay}
\title{Index of Vanclay (1991)}
\usage{
di_vanclay(dbh, area)
}
\description{
Ratio of basal area of trees larger than the subject tree
to basal area of the plot
}
