% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_distance_independent.R
\name{di_wikoff}
\alias{di_wikoff}
\title{Distance-independent Wikoff (1982) index}
\usage{
di_wikoff(dbh, area)
}
\description{
Sum of the basal area of trees larger than the subject tree (m2 ha-1)
}

