% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eucalyptus.R
\docType{data}
\name{eucalyptus}
\alias{eucalyptus}
\title{A amesure plot of Eucalyptus}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 2240 rows and 8 columns.}
\usage{
eucalyptus
}
\description{
A amesure plot of Eucalyptus
}
\examples{

}
\keyword{datasets}

