% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_util.R
\name{join_objtree_compt}
\alias{join_objtree_compt}
\title{Join Object-tree with competitor-tree in the same data.frame}
\usage{
join_objtree_compt(id, x, y, dbh, search, n)
}
\description{
Join Object-tree with competitor-tree in the same data.frame
}

