% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_layout.R
\name{xcoord}
\alias{xcoord}
\title{Local x coordinator}
\usage{
xcoord(x, xspacing = 3, ncol = 8, start = "left-bottom")
}
\description{
Local x coordinator
}
\examples{

#library(ggplot2)
#library(dplyr)
#
#d <- data_frame(a=1:80) \%>\% 
#mutate(
# x = xcoord(a, start = 'left-bottom'),
# y = ycoord(a, start = 'left-bottom')
#) 
#
#ggplot(d, aes(x, y, label = a)) +
# geom_text()

}

