% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_distance_dependent.R
\name{dd_alemdag}
\alias{dd_alemdag}
\title{Distance-dependent Alemdag (1978) index}
\usage{
dd_alemdag(id, x, y, dbh, search = "nearest", n)
}
\description{
Distance-dependent Alemdag (1978) index
}

