% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_distance_dependent.R
\name{dd_daniels}
\alias{dd_daniels}
\title{Distance-dependent Daniels et al. (1986) index}
\usage{
dd_daniels(id, x, y, dbh, search = "nearest", n)
}
\description{
Distance-dependent Daniels et al. (1986) index
}

