% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_distance_dependent.R
\name{dd_martin}
\alias{dd_martin}
\title{Distance-dependent Matin & Ek (1984) index}
\usage{
dd_martin(id, x, y, dbh, search = "nearest", n)
}
\description{
Distance-dependent Matin & Ek (1984) index
}

