% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_distance_independent.R
\name{di_corrona}
\alias{di_corrona}
\title{Distance-independent Corrona & Ferrara (1989)}
\usage{
di_corrona(dbh, area)
}
\description{
Sums up the basal area of neighbours divided by the subject
tree basal area in the plot (ha-1)
}

